+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTableUI.java
@@ -20,11 +20,13 @@ import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.event.FocusEvent;
 import java.awt.event.FocusListener;
+import javax.swing.JCheckBox;
 import javax.swing.JComponent;
 import javax.swing.LookAndFeel;
 import javax.swing.UIManager;
 import javax.swing.plaf.ComponentUI;
 import javax.swing.plaf.basic.BasicTableUI;
+import javax.swing.table.TableCellRenderer;
 import com.formdev.flatlaf.util.UIScale;
 
 /**
@@ -132,6 +134,12 @@ public class FlatTableUI
 			oldIntercellSpacing = table.getIntercellSpacing();
 			table.setIntercellSpacing( intercellSpacing );
 		}
+
+		// checkbox is non-opaque in FlatLaf and therefore would not paint selection
+		// --> make checkbox renderer opaque (but opaque in Metal or Windows LaF)
+		TableCellRenderer booleanRenderer = table.getDefaultRenderer( Boolean.class );
+		if( booleanRenderer instanceof JCheckBox )
+			((JCheckBox)booleanRenderer).setOpaque( true );
 	}
 
 	@Override